feat(be): 평행우주 유사도의 백분위 기반 표시 점수 변환 도입

+++ backend/src/main/java/com/nidus/twinly/common/parallel/ParallelRelationResolver.java
@@ -25,7 +25,7 @@ public class ParallelRelationResolver {
         }
     }
 
-    public ParallelRelationType relationOf(double score) {
+    public ParallelRelationType relationOf(int score) {
         return parallelRelationProperties.relationThresholds().entrySet().stream()
                 .filter(entry -> score >= entry.getValue())
                 .max(Map.Entry.comparingByValue())
